business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetRptrOffset.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetRptrOffset implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+O ");
        messageString.append(command.getParameter("Rptr Offset"));
        messageString.append("\n");
        return messageString;
    }
}
